Fixed bug: Dashboard stuck in Dashboard got stuck in
"loading" state for brand new users. It happened because when the first call to the Firebase.USERS returned the only user id as key,
but empty `value` that we could decode to the User object. As result, the `User.shared.userDataHasBeenLoaded` is never turned to `true`.
Dashboard is waiting for this parameter to update the top icons states. As result, the Dashboard got stuck on the "loading" state.

The complete user's record is created only when we try to update the user_person_type on the Onboarding step.

Make it so:

- add `completion: (() -> ())? = nil` parameter to the User's `save()` method
- on the Onboarding step, when we are saving the `user_person_type` to the Firebase, set `User.shared.userDataHasBeenLoaded=true`